gatk3_merge_summary_metrics: minor fixes for interval summary

+++ shell/variant_calling/gatk3/gatk3_merge_summary_metrics.sh
@@ -143,22 +143,25 @@ done
 
 echo "`${NOW}`INFO $SCRIPT_CODE done"
 
-#transpose interval summary
+if [ -s "$OUTPUT_FILE_TMP" ]; then
 
-echo "`${NOW}`INFO $SCRIPT_CODE transposing interval_summary table..."
+	#transpose interval summary
+	echo "`${NOW}`INFO $SCRIPT_CODE transposing interval_summary table..."
 
-#configure script
-sed -i -e "s/#inputTable/${OUTPUT_FILE_TMP//\//\\/}/" $TMPDIR/transposeTable.R
+	#configure script
+	sed -i -e "s/#inputTable/${OUTPUT_FILE_TMP//\//\\/}/" $TMPDIR/transposeTable.R
 
-#execute script
-R --vanilla < $TMPDIR/transposeTable.R
+	#execute script
+	R --vanilla < $TMPDIR/transposeTable.R
 
-#reformat chromosome coordinates
-perl -i -pe 's/X(.?)\.(\d*?)\.(\d*?)/$1:$2-$3/' $OUTPUT_FILE_TMP.tr
+	#reformat chromosome coordinates
+	perl -i -pe 's/X(.?)\.(\d*?)\.(\d*?)/$1:$2-$3/' $OUTPUT_FILE_TMP.tr
 
-mv $OUTPUT_FILE_TMP.tr $OUTPUT_FILE_TMP
+	mv $OUTPUT_FILE_TMP.tr $OUTPUT_FILE_TMP
 
-cp $OUTPUT_FILE_TMP $OUTPUT_FILE
+	cp $OUTPUT_FILE_TMP $OUTPUT_FILE
+
+fi
 
 #logging
 STATUS=OK
